# openapi.yaml
openapi: 3.0.3
info:
  title: brain-server API
  description: >
    HTTP API for the brain-server local agent memory / semantic search server
    (Axum + rusqlite + sqlite-vec + model2vec). This spec is hand-written from
    the source in src/main.rs and src/handlers/*. Endpoints whose contract could
    not be fully verified from source are marked with `x-status: unverified`.
    Every response carries `X-Api-Version: <semver>`; legacy routes carry a
    `Deprecation` header per the deprecation policy in API_CONTRACT.md.
  version: 0.9.8
  x-api-version: "0.9.8"
servers:
  - url: http://127.0.0.1:8765
    description: Default local server (override BIND_HOST/BIND_PORT)
paths:
...
  /recall:
    post:
      summary: Deterministic end-to-end recall (v0.9.5 structured QueryDoc)
      operationId: recall
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryDoc"
      responses:
        "200":
          description: Recall hits + unified telemetry/provenance envelope
          content:
            application/json:
              schema:
                type: object
                properties:
                  hits:
                    type: array
                    items:
                      $ref: "#/components/schemas/RecallHit"
                  domain: { type: string, nullable: true }
                  domains_searched:
                    type: array
                    nullable: true
                    items: { type: string }
                  telemetry:
                    $ref: "#/components/schemas/SearchTelemetry"
  /ingest:
    post:
      summary: Plugin ingest (contract stub)
      operationId: ingest
      x-status: unverified
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Wire contract per API_CONTRACT.md; body not verified in source.
      responses:
        "200":
          description: Unverified response shape
...
components:
  schemas:
    LexSpec:
      type: object
      description: >
        Structured lexical controls, compiled server-side into a single
        FTS5-quoted MATCH expression (injection-safe). All fields optional;
        empty = fall back to the bare query.
      properties:
        terms:
          type: array
          items: { type: string }
          description: Implicit-AND terms.
        phrases:
          type: array
          items: { type: string }
          description: Quoted phrase matches (ordered token runs).
        exclude:
          type: array
          items: { type: string }
          description: Exclusions, prefixed with FTS5 NOT (-).
        code:
          type: array
          items: { type: string }
          description: Exact identifier / code path, matched verbatim.
    QueryDoc:
      type: object
      description: v0.9.5 structured query document. `v` must be 1 (unknown versions are rejected).
      properties:
        v: { type: integer, default: 1, description: schema version }
        q: { type: string, description: plain-text fallback query }
        query: { type: string, description: recall query (POST /recall) }
        lex: { $ref: "#/components/schemas/LexSpec" }
        vec: { type: string, description: explicit embedding query (overrides q) }
        hyde: { type: string, description: hypothetical answer (overrides vec) }
        intent: { type: string, description: recorded for provenance/telemetry only }
        sources:
          type: array
          items: { type: string }
          description: multi-source OR scope
        source: { type: string, description: legacy single-source equality }
        since: { type: string, description: ISO-8601 / YYYY-MM-DD HH:MM:SS lower bound }
        domain: { type: string, description: domain isolation label }
        k: { type: integer, description: result count (GET /search) }
        limit: { type: integer, description: result count (POST /recall) }
        profile: { type: string, description: retrieval profile hint (passthrough) }
        explain: { type: boolean, description: "GET /search query_plan toggle" }
        provenance: { type: boolean, description: "POST /recall telemetry/provenance toggle" }
        include_flagged: { type: boolean, description: "v0.9.7 Guard: include quarantined chunks" }
        as_of: { type: string, nullable: true, description: "v0.9.8 Evidence: point-in-time recall (RFC3339)" }
        evidence: { type: boolean, description: "v0.9.8 Evidence: include structured Evidence on every hit" }
    Evidence:
      type: object
      description: >-
        Faithful, bounded evidence for a hit (verbatim snippet + span + source
        link). v0.9.8 adds time semantics, lifecycle, and typed links.
      properties:
        text: { type: string, description: verbatim substring of content }
        line_start: { type: integer, nullable: true }
        line_end: { type: integer, nullable: true }
        heading_path: { type: string, nullable: true }
        source_uri: { type: string, nullable: true }
        revision_id: { type: integer, nullable: true }
        valid_from: { type: string, nullable: true, description: when the fact became true (RFC3339) }
        valid_to: { type: string, nullable: true, description: when the fact ceased to be true; null = current }
        observed_at: { type: string, nullable: true, description: when brain-server learned the fact (RFC3339) }
        authority: { type: number, nullable: true, description: "source-authority tie-breaker (0..1)" }
        lifecycle:
          type: string
          nullable: true
          enum: [current, historical, superseded]
          description: current vs historical vs superseded; omitted when current
        links:
          type: array
          items:
            type: object
            properties:
              to_chunk: { type: integer }
              kind: { type: string, enum: [supports, supersedes, contradicts, references, derived_from] }
        highlights:
          type: array
          items:
            type: array
            items: { type: integer }
            description: "[start,end) byte offsets within `text`"
...
    RecallHit:
      type: object
      properties:
        id: { type: integer }
        title: { type: string, nullable: true }
        content: { type: string }
        snippet: { type: string, description: bounded verbatim preview }
        score: { type: number }
        domain: { type: string, nullable: true }
        source:
          type: string
          enum: [vector, fts, both, graph]
          nullable: true
        evidence:
          $ref: "#/components/schemas/Evidence"
        conflict:
          type: boolean
          nullable: true
          description: "v0.9.8: true when this chunk has a contradicts/supersedes link to a current chunk"
        provenance:
          type: object
          nullable: true
          properties:
            vector_rank: { type: integer, nullable: true }
            fts_rank: { type: integer, nullable: true }
            fused_score: { type: number, nullable: true }
            rerank_score: { type: number, nullable: true }
            rerank_truncated: { type: boolean }
            prf_expanded: { type: boolean }
...
    SearchTelemetry:
      type: object
      description: Unified retrieval telemetry (POST /recall provenance=true).
      properties:
        embed_ms: { type: number }
        vector_ms: { type: number }
        fts_ms: { type: number }
        fusion_ms: { type: number }
        prf_ms: { type: number }
        rerank_ms: { type: number }
        vec_candidates: { type: integer }
        fts_candidates: { type: integer }
        fused_count: { type: integer }
        rrf_k: { type: integer }
        embedding_query: { type: string, nullable: true }
        intent: { type: string, nullable: true }
        retrieval_ms_vec: { type: number }
        retrieval_ms_fts: { type: number }
        confidence: { type: number }
        recommendation: { type: string, nullable: true }
...